+++ js/ckeditor5_plugins/buttongroup/src/buttongroupui.js
@@ -1,11 +1,10 @@
 import { Plugin } from 'ckeditor5/src/core';
 import { ButtonView, ContextualBalloon, clickOutsideHandler } from 'ckeditor5/src/ui';
-import FormView from './buttonview';
+import BGFormView from './buttongroupview';
 
-import getRangeText from './buttonutils.js';
-import icon from '../../../../icons/arrow-right-to-bracket-solid.svg';
+import icon from '../../../../icons/arrows-turn-right-solid.svg';
 
-export default class ButtonUI extends Plugin {
+export default class ButtonGroupUI extends Plugin {
 	static get requires() {
 		return [ ContextualBalloon ];
 	}
@@ -13,7 +12,7 @@ export default class ButtonUI extends Plugin {
 	init() {
 		const editor = this.editor;
 		const componentFactory = editor.ui.componentFactory;
-		const insertButtonCommand = editor.commands.get('addButton')
+		const insertButtonGroupCommand = editor.commands.get('addButtonGroup')
 		const viewDocument = editor.editing.view.document;
 
         // Create the balloon and the form view.
@@ -22,39 +21,39 @@ export default class ButtonUI extends Plugin {
 		this.buttonView = null;
 
 		// This will register the Button button to the toolbar
-		componentFactory.add( 'button', (locale) => {
+		componentFactory.add( 'buttonGroup', (locale) => {
 			const button = new ButtonView(locale);
 
-			button.label = 'Button';
+			button.label = 'Button Group';
 			button.icon = icon;
 			button.tooltip = true;
 			button.withText = false;
 			button.isToggleable = true;
 			// Show the UI on button click.
 			this.listenTo( button, 'execute', () => {
-				this._showUI(insertButtonCommand.existingButtonSelected)
+				this._showUI(insertButtonGroupCommand.existingButtonGroupSelected)
 			} );
 
 			this.buttonView = button;
 
 			// Show the on/off in Toolbar if a button is already selected.
 			const updateButtonState = () => {
-				const linkButtonSelected = insertButtonCommand.existingButtonSelected;
+				const linkButtonSelected = insertButtonGroupCommand.existingButtonGroupSelected;
 				button.isOn = !!linkButtonSelected;
 			};
 			  
 			 // Listen for changes in the linkButton selection.
-			this.listenTo(insertButtonCommand, 'change:value', updateButtonState);
-			this.listenTo(insertButtonCommand, 'change:existingButtonSelected', updateButtonState);
+			this.listenTo(insertButtonGroupCommand, 'change:value', updateButtonState);
+			this.listenTo(insertButtonGroupCommand, 'change:existingButtonSelected', updateButtonState);
 			
 			// Shows the UI on click of a button widget.
 			this.listenTo(viewDocument, 'click', () => {
-				if (insertButtonCommand.existingButtonSelected)
-					this._showUI(insertButtonCommand.existingButtonSelected);
+				if (insertButtonGroupCommand.existingButtonGroupSelected)
+					this._showUI(insertButtonGroupCommand.existingButtonGroupSelected);
 			});
 
 			// Bind the state of the button to the command.
-			button.bind('isOn', 'isEnabled').to(insertButtonCommand, 'value', 'isEnabled');
+			button.bind('isOn', 'isEnabled').to(insertButtonGroupCommand, 'value', 'isEnabled');
 
 			return button;
 		} );
@@ -63,20 +62,17 @@ export default class ButtonUI extends Plugin {
 	_createFormView(locale) {
 		const editor = this.editor;
 		const componentFactory = editor.ui.componentFactory;
-		const formView = new FormView( locale, componentFactory);
+		const formView = new BGFormView( locale, componentFactory);
 
 
 		// Execute the command after clicking the "Save" button.
 		this.listenTo( formView, 'submit', () => {
 			// Grab values from the Form and title input fields.
 			const value = {
-				href: formView.linkInputView.fieldView.element.value,
 				color: formView.color,
 				size: formView.size,
-				style: formView.style
-
 			};
-			editor.execute( 'addButton', value );
+			editor.execute( 'addButtonGroup', value );
 
             // Hide the form view after submit.
 			this._hideUI();
@@ -107,42 +103,35 @@ export default class ButtonUI extends Plugin {
 	_showUI(selectedButton) {
 		const selection = this.editor.model.document.selection;
 		this.buttonView.isOn = true;
+	
 		// Check the value of the command.
-		const commandValue = this.editor.commands.get( 'addButton' ).value;
-
-		this._balloon.add( {
-			view: this.formView,
-			position: this._getBalloonPositionData()
-		} );
-
-		if (selectedButton) {
-			const size = selectedButton.getAttribute('linkButtonSize');
-			const color = selectedButton.getAttribute('linkButtonColor');
-			const style = selectedButton.getAttribute('linkButtonStyle');
-			const href = selectedButton.getAttribute('linkButtonHref');
-		
-			this.formView.color = color;
-			this.formView.style = style;
-			this.formView.size = size;
-		
-			this.formView.linkInputView.fieldView.value = href;
-			this.formView.linkInputView.fieldView.element.value = href; // Update the input field value
-			this.formView.linkInputView.fieldView.set('value', href); // Update the input field value (alternative method)
+		const commandValue = this.editor.commands.get('addButtonGroup').value;
+	
+		// Check if the balloon is already open for the selectedButton
+		if (!this._balloon.visibleView || this._balloon.visibleView !== this.formView) {
+			this._balloon.add({
+				view: this.formView,
+				position: this._getBalloonPositionData(),
+			});
 		}
-
-		// Disable the input when the selection is not collapsed.
-		// this.formView.linkInputView.isEnabled = selection.getFirstRange().isCollapsed;
-
+	
+		if (selectedButton) {
+			const size = selectedButton.getAttribute('buttonGroupSize');
+			const color = selectedButton.getAttribute('buttonGroupColor');
+	
+			this.formView.color = color;
+			this.formView.size = size;
+		}
+	
 		// Fill the form using the state (value) of the command.
-		if ( commandValue ) {
-			this.formView.linkInputView.fieldView.value = commandValue.link;
+		if (commandValue) {
 			this.formView.colorDropdown.fieldView.value = commandValue.color;
 			this.formView.sizeDropdown.fieldView.value = commandValue.size;
-			this.formView.styleDropdown.fieldView.value = commandValue.style
 		}
-
+	
 		this.formView.focus();
 	}
+	
 
 	_hideUI() {
 		// Clear the input field values and reset the form.