+++ js/ckeditor5_plugins/buttongroup/src/index.js
@@ -1,0 +1,5 @@
+import ButtonGroup from "./buttongroup";
+
+export default {
+  ButtonGroup
+};